+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - **BC Break**: Remove bridges for React Native and validate.js.
 - Added `errors`, `lastErrors`, `hasErrors` and `hasLastErrors` properties to the form object.
 - Added `revalidateFields` to the form option, to opt-in which fields should revalidate on change.
+- Added `preValidateTransform` and `postValidateTransform` to the form option
 - Added `isDirty` to the field prop, to indicate whether the field has seen a change even since the last submission.
 - `onSubmit` can now return a `boolen` or `Promise<boolean>`, to indicate whether submission was successful (useful for code awaiting `handleSubmit`).
 - Validation functions can now return an empty array for a field to indicate no errors, rather than only `undefined`.